requirements.txt
asgiref==3.8.1
Django==5.1.5
django-jazzmin==3.0.1
django-tinymce==4.1.0
pillow==11.1.0
sqlparse==0.5.3
tzdata==2024.2
